Commit 8: - code clean-up - added creation date and last edited to user details - added confirmation when deleting the user

+++ src/main/java/pl/coderslab/users/UserDelete.java
@@ -11,10 +11,12 @@ import java.io.IOException;
 
 @WebServlet(name = "UserDelete", urlPatterns = "/user/delete")
 public class UserDelete extends HttpServlet {
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
-        request.setCharacterEncoding("UTF-8");
+    private static final String URL_DELETE = "/users/delete.jsp";
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
         String stringId = request.getParameter("id");
 
         int id;
@@ -27,8 +29,40 @@ public class UserDelete extends HttpServlet {
         }
 
         UserDao userDao = new UserDao();
-        userDao.delete(id);
+        User user = userDao.read(id);
 
-        response.sendRedirect("/user/list");
+        request.setAttribute("user", user);
+        request.setAttribute("deleted", false);
+        getServletContext().getRequestDispatcher(URL_DELETE).forward(request, response);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+        request.setCharacterEncoding("UTF-8");
+        String stringId = request.getParameter("id");
+        String stringIsConfirmed = request.getParameter("isConfirmed");
+
+
+        boolean isConfirmed = Boolean.parseBoolean(stringIsConfirmed);
+
+        int id;
+        try {
+            id = Integer.parseInt(stringId);
+        } catch (NumberFormatException e) {
+            request.setAttribute("userNotFound", true);
+            getServletContext().getRequestDispatcher("/users/show.jsp").forward(request, response);
+            return;
+        }
+
+        if (isConfirmed) {
+            UserDao userDao = new UserDao();
+            userDao.delete(id);
+            request.setAttribute("deleted", true);
+            getServletContext().getRequestDispatcher(URL_DELETE).forward(request, response);
+        } else {
+            response.sendRedirect("/user/list");
+        }
+
     }
 }
